Ready to implement inv kin calculations. Attempting to fix rviz rip

+++ scripts/inverse_kinematics_client.py
@@ -21,9 +21,9 @@ def usage():
 if __name__ == "__main__":
     
     if len(sys.argv) == 4:
-        x = int(sys.argv[1])
-        y = int(sys.argv[2])
-        z = int(sys.argv[3])
+        x = float(sys.argv[1])
+        y = float(sys.argv[2])
+        z = float(sys.argv[3])
     else:
         print(usage())
         sys.exit(1)
